fix warning
git-svn-id: https://svn.salilab.org/imp/trunk@11783 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/multifit/include/anchors_reader.h
@@ -31,7 +31,7 @@ class IMPMULTIFITEXPORT AnchorsData {
     bool broken=false;
     for (unsigned int op=0;op<orig.points_.size();op++) {
       for (unsigned int np=0;np<these_indexes.size();np++) {
-        if (these_indexes[np]==op) {
+        if (these_indexes[np]==static_cast<int>(op)) {
           in_subset.push_back(1);
           broken=true;
           if (keep) {
